+++ src/Field.php
@@ -95,11 +95,23 @@ class Field
     }
 
     /**
+     * Get the field parent key.
+     *
+     * @return string
+     */
+    public function getParentKey(): string
+    {
+        return $this->parentKey;
+    }
+
+    /**
      * Set the field parent key.
+     *
+     * @param string $parentKey
      *
      * @return void
      */
-    public function setParentKey(string $parentKey)
+    public function setParentKey(string $parentKey): void
     {
         $this->parentKey = $parentKey;
     }
@@ -114,7 +126,7 @@ class Field
         $conditionals = [];
 
         foreach ($this->settings['conditional_logic'] as $rules) {
-            $conditional = new Conditional($rules, $this->getKey());
+            $conditional = new Conditional($rules, $this->getParentKey());
 
             $conditionals[] = $conditional->toArray();
         }