+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.0.0 (upcoming)
 
 - Added hashed field, group and layout keys
+- Fixed field keys in conditional logic
 
 ## 3.1.0 (released 2017-11-30)